feat: add file, git tool for test
